+++ R/grouped-df.r
@@ -196,13 +196,43 @@ cbind.grouped_df <- function(...) {
   bind_cols(...)
 }
 
+#' Select grouping variables
+#'
+#' This selection helpers matches grouping variables. It can be used
+#' in [select()] or [vars()][scoped] selections.
+#'
+#' @inheritParams tidyselect::select_helpers
+#' @seealso [groups()] and [group_vars()] for retrieving the grouping
+#'   variables outside selection contexts.
+#'
+#' @examples
+#' gdf <- iris %>% group_by(Species)
+#'
+#' # Select the grouping variables:
+#' gdf %>% select(group_cols())
+#'
+#' # Remove the grouping variables from mutate selections:
+#' gdf %>% mutate_at(vars(-group_cols()), `/`, 100)
+#' @export
+group_cols <- function(vars = peek_vars()) {
+  if (is_sel_vars(vars)) {
+    matches <- match(vars %@% groups, vars)
+    if (anyNA(matches)) {
+      abort("Can't find the grouping variables")
+    }
+    matches
+  } else {
+    int()
+  }
+}
+
 # One-table verbs --------------------------------------------------------------
 
 # see arrange.r for arrange.grouped_df
 
 .select_grouped_df <- function(.data, ..., notify = TRUE) {
   # Pass via splicing to avoid matching vars_select() arguments
-  vars <- tidyselect::vars_select(names(.data), !!!quos(...))
+  vars <- tidyselect::vars_select(sel_vars(.data), !!!quos(...))
   vars <- ensure_group_vars(vars, .data, notify = notify)
   select_impl(.data, vars)
 }